adicionando projeto de JSF

+++ Programação/Java/Java EE/Java e JSF I - Sua Aplicação Web com JSF2/livraria/src/br/com/caelum/livraria/dao/DAO.java
@@ -62,7 +62,7 @@ public class DAO<T> {
 		return lista;
 	}
 
-	public T buscaPorId(Integer id) {
+	public T buscaPorId(Long id) {
 		EntityManager em = new JPAUtil().getEntityManager();
 		T instancia = em.find(classe, id);
 		em.close();
